requirements.txt
Flask==1.1.4
Flask==2.2.2
joblib==1.2.0
matplotlib==3.0.0
matplotlib==3.7.2
numpy==1.15.2
numpy==1.25.2
pandas==2.0.3
Pillow==9.4.0
Pillow==10.3.0
plotly==5.22.0
scikit_learn==0.20.0
scikit_learn==1.3.0
seaborn==0.13.2
tensorflow==2.16.1
tqdm==4.63.0
tqdm==4.65.0
webcolors==1.13
